+++ docker-java-api/src/main/java/com/github/dockerjava/api/model/Task.java
@@ -97,9 +97,7 @@ public class Task extends DockerObject implements Serializable {
      * @return name
      **/
 
-    public String getName() {
-        return name;
-    }
+
 
     /**
      * User-defined key/value metadata.
