+++ docker-java-api/src/main/java/com/github/dockerjava/api/model/Ulimit.java
@@ -43,9 +43,7 @@ public class Ulimit extends DockerObject implements Serializable {
         this.hard = hard;
     }
 
-    public String getName() {
-        return name;
-    }
+
 
     @Deprecated
     @JsonIgnore
